MBX-2706 New functional

+++ Mindbox/InAppMessages/Models/Config/Inapp/InappModel/InappForm/InappFormVariant/InappFormVariantContent/ContentElement/ContentElementSize/ContentElementSize.swift
@@ -10,6 +10,6 @@ import Foundation
 
 struct ContentElementSize: Decodable, Equatable {
     let kind: ContentElementSizeKind
-    let width: Decimal
-    let height: Decimal
+    let width: Double
+    let height: Double
 }
